use ni for generating install commands
