+++ packages/internals/src/index.ts
@@ -30,6 +30,7 @@ export { getQueryEngineProtocol, type QueryEngineProtocol } from './client/getQu
 export { credentialsToUri, protocolToConnectorType, uriToCredentials } from './convertCredentials'
 export * from './engine-commands'
 export { Generator } from './Generator'
+export type { GeneratorPaths } from './get-generators/generatorResolvers/generatorResolvers'
 export type { ProviderAliases } from './get-generators/getGenerators'
 export { getGenerator, getGenerators } from './get-generators/getGenerators'
 export { getPackedPackage } from './getPackedPackage'
@@ -39,7 +40,6 @@ export type { MigrateEngineLogLine } from './migrateEngineCommands'
 export { canConnectToDatabase, createDatabase, dropDatabase } from './migrateEngineCommands'
 export { MigrateEngineExitCode } from './migrateEngineCommands'
 export { ErrorArea, isRustPanic, RustPanic } from './panic'
-export type { GeneratorPaths } from './predefinedGeneratorResolvers'
 export { BinaryType } from './resolveBinary'
 export { engineEnvVarMap, resolveBinary } from './resolveBinary'
 export { sendPanic } from './sendPanic'
